Struct aws_smithy_types::body::SdkBody

pub struct SdkBody { /* private fields */ }

SdkBody type

This is the Body used for dispatching all HTTP Requests. For handling responses, the type of the body will be controlled by the HTTP stack.

Implementations

impl SdkBody

pub fn from_body_0_4<T, E>(body: T) -> Selfwhere T: Body<Data = Bytes, Error = E> + Send + Sync + 'static, E: Into<Error> + 'static,

Construct an SdkBody from a type that implements http_body_0_4::Body<Data = Bytes>.

Note: This is only available with http-body-0-4-x enabled.

impl SdkBody

pub fn retryable(f: impl Fn() -> SdkBody + Send + Sync + 'static) -> Self

Construct an explicitly retryable SDK body

Note: This is probably not what you want

All bodies constructed from in-memory data (String, Vec<u8>, Bytes, etc.) will be retryable out of the box. If you want to read data from a file, you should use ByteStream::from_path. This function is only necessary when you need to enable retries for your own streaming container.

pub fn taken() -> Self

When an SdkBody is read, the inner data must be consumed. In order to do this, the SdkBody is swapped with a “taken” body. This “taken” body cannot be read but aids in debugging.

pub fn empty() -> Self

Create an empty SdkBody for requests and responses that don’t transfer any data in the body.

pub fn bytes(&self) -> Option<&[u8]>

If possible, return a reference to this body as &[u8]

If this SdkBody is NOT streaming, this will return the byte slab If this SdkBody is streaming, this will return None

pub fn try_clone(&self) -> Option<Self>

Attempt to clone this SdkBody. This will fail if the inner data is not cloneable, such as when it is a single-use stream that can’t be recreated.

pub fn content_length(&self) -> Option<u64>

Return the length, in bytes, of this SdkBody. If this returns None, then the body does not have a known length.

pub fn map( self, f: impl Fn(SdkBody) -> SdkBody + Sync + Send + 'static ) -> SdkBody

Given a function to modify an SdkBody, run that function against this SdkBody before returning the result.

pub fn map_preserve_contents( self, f: impl Fn(SdkBody) -> SdkBody + Sync + Send + 'static ) -> SdkBody

Update this SdkBody with map. This function MUST NOT alter the data of the body.

This function is useful for adding metadata like progress tracking to an SdkBody that does not alter the actual byte data. If your mapper alters the contents of the body, use SdkBody::map instead.