aws_smithy_types::byte_stream

Struct AggregatedBytes

pub struct AggregatedBytes(/* private fields */);

Non-contiguous Binary Data Storage

When data is read from the network, it is read in a sequence of chunks that are not in contiguous memory. AggregatedBytes provides a view of this data via impl Buf or it can be copied into contiguous storage with .into_bytes().

Implementations

impl AggregatedBytes

pub fn into_bytes(self) -> Bytes

Convert this buffer into Bytes.

Why does this consume self?

Technically, copy_to_bytes can be called without ownership of self. However, since this mutates the underlying buffer such that no data is remaining, it is more misuse resistant to prevent the caller from attempting to reread the buffer.

If the caller only holds a mutable reference, they may use copy_to_bytes directly on AggregatedBytes.

pub fn into_segments(self) -> impl Iterator<Item = Bytes>

Convert this buffer into an Iterator of underlying non-contiguous segments of Bytes

pub fn to_vec(self) -> Vec<u8>

Convert this buffer into a Vec<u8>