Struct aws_smithy_types::config_bag::Layer

pub struct Layer { /* private fields */ }

A named layer comprising a config bag

See the module docs for more documentation.

Implementations

impl Layer

pub fn is_empty(&self) -> bool

Returns true if this layer is empty.

pub fn freeze(self) -> FrozenLayer

Converts this layer into a frozen layer that can no longer be mutated.

pub fn new(name: impl Into<Cow<'static, str>>) -> Self

Create a new Layer with a given name

pub fn with_name(self, name: impl Into<Cow<'static, str>>) -> Self

Changes the name of this layer.

pub fn load<T: Storable>(&self) -> <T::Storer as Store>::ReturnedType<'_>

Load a storable item from the bag

pub fn unset<T: Send + Sync + Debug + 'static>(&mut self) -> &mut Self

Remove T from this bag

pub fn store_put<T>(&mut self, item: T) -> &mut Self
where T: Storable<Storer = StoreReplace<T>>,

Stores item of type T into the config bag, overriding a previous value of the same type

pub fn store_or_unset<T>(&mut self, item: Option<T>) -> &mut Self
where T: Storable<Storer = StoreReplace<T>>,

Stores item of type T into the config bag, overriding a previous value of the same type, or unsets it by passing a None

pub fn store_append<T>(&mut self, item: T) -> &mut Self
where T: Storable<Storer = StoreAppend<T>>,

This can only be used for types that use StoreAppend

use aws_smithy_types::config_bag::{ConfigBag, Layer, Storable, StoreAppend, StoreReplace};
let mut layer_1 = Layer::new("example");
#[derive(Debug, PartialEq, Eq)]
struct Interceptor(&'static str);
impl Storable for Interceptor {
    type Storer = StoreAppend<Interceptor>;
}

layer_1.store_append(Interceptor("321"));
layer_1.store_append(Interceptor("654"));

let mut layer_2 = Layer::new("second layer");
layer_2.store_append(Interceptor("987"));

let bag = ConfigBag::of_layers(vec![layer_1, layer_2]);

assert_eq!(
    bag.load::<Interceptor>().collect::<Vec<_>>(),
    vec![&Interceptor("987"), &Interceptor("654"), &Interceptor("321")]
);

pub fn clear<T>(&mut self)
where T: Storable<Storer = StoreAppend<T>>,

Clears the value of type T from the config bag

This internally marks the item of type T as cleared as opposed to wiping it out from the config bag.