aws_smithy_types::config_bag

Trait Storable

pub trait Storable:
    Send
    + Sync
    + Debug
    + 'static {
    type Storer: Store;
}

Trait that marks the implementing types as able to be stored in the config bag

See the module docs for more documentation.

Required Associated Types

type Storer: Store

Specify how an item is stored in the config bag, e.g. StoreReplace and StoreAppend