aws_smithy_types::error::metadata

Struct ErrorMetadata

pub struct ErrorMetadata { /* private fields */ }

Generic Error type

For many services, Errors are modeled. However, many services only partially model errors or donโt model errors at all. In these cases, the SDK will return this generic error type to expose the code, message and request_id.

Implementations

impl ErrorMetadata

pub fn code(&self) -> Option<&str>

Returns the error code.

pub fn message(&self) -> Option<&str>

Returns the error message.

pub fn extra(&self, key: &'static str) -> Option<&str>

Returns additional information about the error if itโs present.

pub fn builder() -> Builder

Creates an Error builder.

pub fn into_builder(self) -> Builder

Converts an Error into a builder.