Module aws_smithy_types::retry

This module defines types that describe when to retry given a response.

Structs

RetryConfig

Retry configuration for requests.

RetryConfigBuilder

Builder for RetryConfig.

RetryModeParseErr

Failure to parse a RetryMode from string.

Enums

ErrorKind

Type of error that occurred when making a request.

RetryConfigErr

Failure to parse retry config from profile file or environment variable.

RetryKind

RetryKind describes how a request MAY be retried for a given response

RetryMode

Specifies how failed requests should be retried.

Traits

ProvideErrorKind

Trait that provides an ErrorKind and an error code.