Crate aws_smithy_types

Protocol-agnostic types for smithy-rs.

Re-exports

pub use date_time::DateTime;

Modules

base64

A thin wrapper over base64-simd

body

Types for representing the body of an HTTP request or response

byte_stream

ByteStream Abstractions

checksum_config

Types that allow users to indicate their preferences for checksum calculation and validation

config_bag

A typemap for storing configuration. Layers and layered bags of configuration data.

date_time

DateTime type for representing Smithy timestamps.

endpoint

Smithy Endpoint Types

error

Errors for Smithy codegen

event_stream

Types relevant to event stream serialization/deserialization

primitive

Utilities for formatting and parsing primitives

retry

This module defines types that describe when to retry given a response.

str_bytes

UTF-8 string byte buffer representation with validation amortization.

timeout

This module defines types that describe timeouts that can be applied to various stages of the Smithy networking stack.

type_erasure

Utilities for type erasure.

Structs

Blob

Binary Blob Type

Enums

Document

Document Type

Number

A number type that implements Javascript / JSON semantics, modeled on serde_json: https://docs.serde.rs/src/serde_json/number.rs.html#20-22