Crate aws_smithy_types

Protocol-agnostic types for smithy-rs.

Re-exports

pub use crate::date_time::DateTime;

pub use error::Error;

Modules

base64

A correct, small, but not especially fast base64 implementation

date_time

DateTime type for representing Smithy timestamps.

error

Generic errors for Smithy codegen

primitive

Utilities for formatting and parsing primitives

retry

This module defines types that describe when to retry given a response.

timeout

This module defines types that describe timeouts that can be applied to various stages of the Smithy networking stack.

tristate

Contains TriState definition and impls

Structs

Blob

Binary Blob Type

Enums

Document

Document Type

Number

A number type that implements Javascript / JSON semantics, modeled on serde_json: https://docs.serde.rs/src/serde_json/number.rs.html#20-22