Trait aws_smithy_types::config_bag::Store

pub trait Store: Sized + Send + Sync + 'static {
    type ReturnedType<'a>: Send + Sync;
    type StoredType: Send + Sync + Debug;

    // Required method
    fn merge_iter(iter: ItemIter<'_, Self>) -> Self::ReturnedType<'_>;
}

Trait defining how types can be stored and loaded from the config bag

Required Associated Types

type ReturnedType<'a>: Send + Sync

Denote the returned type when loaded from the config bag

type StoredType: Send + Sync + Debug

Denote the stored type when stored into the config bag

Required Methods

fn merge_iter(iter: ItemIter<'_, Self>) -> Self::ReturnedType<'_>

Create a returned type from an iterable of items

Object Safety

This trait is not object safe.